[deps] update gl-native to beta.15 and common to beta.11, adopt break… (#120)
* [deps] update gl-native to beta.15 and common to beta.11, adopt breaking changes

* Fix breaking changes related to text-field and GeoJsonSource data property.

+++ sdk/src/main/java/com/mapbox/maps/NativeMapImpl.kt
@@ -140,7 +140,7 @@ internal class NativeMapImpl(private val map: MapInterface) :
     return map.cameraForGeometry(geometry, edgeInsets, zoom, pitch)
   }
 
-  override fun getElevation(point: Point): Float? {
+  override fun getElevation(point: Point): Double? {
     return map.getElevation(point)
   }
